% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Rrs.SAS.R
\name{compute.Rrs.SAS}
\alias{compute.Rrs.SAS}
\title{Compute the remote sensing reflectance from HyperSAS data}
\arguments{
\item{SAS}{is a list return by \code{\link{read.horc.L2.SAS}}.}

\item{tilt.max}{is the maximum tilt tolerance for the SAS.
A default value is 3 degrees is used otherwise.}

\item{quantile.prob}{is a value (betwwen 0.25 to 1) for the maximum quantile probability
for which the surface radiance (Lt) values will be discarded.
The quantile probability is the value at which the probability
of the random variable being less than or equal to this value.
For example, a value of 0.5 means that every Lt value above the
50\% quantile probability will be discarded, while a value of 0.8 means
that only the values above the 80\% quantile probability will be discarded.
The rational is to eliminate outliers resulting from sun glitters, foam, sea ice, etc.
The default value is 0.5 to eliminate Lt value above the 50\% quantile probalibity.}

\item{windspeed}{is the wind speed in m/s.
A default value of 5 m/s is used otherwise.}

\item{thetaV}{is the viewing sensor zenith angle in degree.
A default value of 40 degrees is used otherwise.}

\item{Dphi}{is the azimuthal difference between the sun and the sensor view angle.
A default value of 90 degrees is used otherwise.}

\item{NIR.CORRECTION}{is the method used to eliminate residual "white reflectance".
Such residual is frequent under windy condisitons at sea.
Those may be due to sun glint, foam, ocean spray, etc.
The "NULL" correction subtracts the residual Rrs value at 800 nm.
The "SIMILARITY" correction is prefer in turbid waters. It is based on
the similarity spectrum (see Ruddick et al. L&O 2005) in the NIR. Here it
assumes a constant Rrs(780)/Rrs(720) ratio of 2.35.}

\item{use.COMPASS}{is a logical parameter indicating whether or not the azimuth difference
between the sun and the view geometry is calculated using the SAS compass data.
When FALSE, the azimuth angle difference is taken
from the cast.info.dat file. NOTE: The Default is FALSE because
compass usually doesn't work on ship.}
}
\description{
This is the main function of the package. It processes the SAS data
for a given data acquisition file.
}
\details{
This functions is the main part of the HyperSAS data processing.

First it computes the tilt from the Pitch and Roll data recorded by the SAS
and stored in SAS$Anc data frame. Then tilt values areassociated to
each irradiance (Ed), sky radiance (Li) and surface radiance (Lt)
frame found whitin the file (each of them have their
specific time integration).
Second, the Lt frames are discarded if
1) the tilt is greater than the specified tilt.max;
2) the Lt value around 490 nm is greater than the specified quantile.prob
(to remove upper outliers); and
3) the Lt value around 490 nm is below the 10\% quantile probalility
(to remove lower outliers).

Second, for each valid Lt frame, a corresponding value of Ed and Li is interpolated
using a spline function. Mean and standard devivation of valid Lt
and interpolated Ed and Li spectra are computed.

Third, a spectral interpolation is performed on the average
Ed, Li and Lt spectra from 380 to 800 nm every 5 nm.
This step is done using a \code{\link{loess}} function with a span of 0.05.

Forth, the sky reflectance (Li/Ed) and surface (Lt/Ed) are computed and smoothed using a
\code{\link{loess}} function with a span of 0.75 and 0.1, respectively. This
step is necessary to avoid spectral artefact resulting from the individual spectral
response of each sensor.

Fift, the specular reflectance of the air-sea interface is determined
(i.e., rho_{sky}).
This is based on Ruddick et al. L&0 2005 and Mobley, App. Opt. 1999.
Under clear sky, i.e. when the sky reflectance at 750 nm is below 5\%,
rho_{sky} is interpolated from a Look-Up-Table provided by Mobley.
The rho_{sky} LUT has 4 dimensions for thetaS, thetaV, Dphi
and Windspeed. The current table is not spectrally dependent
(so strictly valid for 550 nm). Spectral dependency is therefore
not taken into account, which may be significant
(see Lee et al, Opt. Exp. 2010)
}
\seealso{
\code{\link{process.HyperSAS}} and \code{\link{read.hocr.SAS}}
}
\author{
Simon Bélanger
}
