% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.HyperSAS.report.R
\name{create.HyperSAS.report}
\alias{create.HyperSAS.report}
\title{Create a PDF report for the HyperSAS data collection for a station or a transit}
\usage{
create.HyperSAS.report(dirdat, TYPE = "STATION", DIAGNOSTIC.PLOTS = FALSE)
}
\arguments{
\item{dirdat}{is the current directory path contaning the files to process.}

\item{TYPE}{is either equal to "STATION" or "TRANSIT" depending
whether the folder gather replicate Rrs of the same station, or the data
collected during the course of a ship transit. If the function is
called from \code{\link{HyperSAS.go}}, the TYPE will be read in the file named
directories.for.HyperSAS.dat. Otherwise the default is "STATION".}

\item{DIAGNOSTIC.PLOTS}{is a logical parameter indicating whether or not diagnostic plots are saved in PNG format.
It is used only when TYPE="TRANSIT".
The plots shows the temporal variability in Ed and tilt,
as well as the spectra for each Ed, Li and Lt recorded during the data acquisition.
Default is FALSE.}
}
\description{
A PDF report is generated using knitr package facilities and a *.Rnw
template. It provides an efficent mean to visualise the data.
It can be called from the higher level function \code{\link{HyperSAS.go}}, or
run in the command line as along as SAS.raw.RData and RRS.RData have been
generated by \code{\link{process.HyperSAS}}.
}
\details{
There is two templates located in the package data folder:
"HyperSAS_Station_TEMPLATE.Rnw" and "HyperSAS_Transit_TEMPLATE.Rnw".
This function will extract a number of informations from the current
data directory and RData files. Next it uses the brew package to update the Rnw files.
The Rnw files are converted into PDF using knitr package.
}
\seealso{
\code{\link{process.HyperSAS}}, \code{\link{HyperSAS.go}}
}
\author{
Simon Bélanger
}
