% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.SAS.rhow.DB.R
\name{generate.SAS.rhow.DB}
\alias{generate.SAS.rhow.DB}
\title{Generate a data base for rhow (pi*Rrs) from the HyperSAS}
\arguments{
\item{path}{is the path where the file directories.for.HyperSAS.dat containing the data folders
to merge in the data base.}

\item{mission}{is a string for the name of the mission. It will be used for the file names of the output.}

\item{wave.range}{is a vector of two integers indicating the wavelength range to output in the database.
The default is wave.range=c(350,810) to put the data from 350 nm to 810 nm.}
}
\value{
It returns a list object named SAS.DB containing a matrix of rhow (rhow.m) and vectors for
StationID, data, lat, lon sunzen, windspeed, rhow.Method.

The object SAS.DB is saved in RData format. The data are also saved in ASCII (.dat with comma separator)
and a figure showing the measured rho_w spectra of the data base is produced.
}
\description{
Generate a data base for rhow (pi*Rrs) from the HyperSAS
}
\author{
Simon Bélanger
}
