% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAA.v5.R
\name{QAA.v5}
\alias{QAA.v5}
\title{Quasi-Analytical Algorithm from Lee et al. App. Opt. (2002)}
\arguments{
\item{waves}{is a vector of wavelenghts}

\item{Rrs}{is the remote sensing reflectance having the same
length as waves.}
}
\description{
This is the version 5
}
\author{
Simon Bélanger
}
