% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QWIP.Rrs.R
\name{QWIP.Rrs}
\alias{QWIP.Rrs}
\title{QWIP: A Quantitative Metric for Quality Control of Aquatic Reflectance
Spectral Shape Using the Apparent Visible Wavelength (Dierssen et al., FRS, 2022)}
\arguments{
\item{RRS}{A list generated by \code{\link{compute.Rrs.SAS}}}
}
\value{
AVW, NDI, QWIP and QWIP.score
}
\description{
This function calculates the Quality Water Index Polynomial (QWIP)
based on Diesrssen et al 2022 paper.
}
\author{
Simon Bélanger
}
