% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hocr.L2.SAS.R
\name{read.hocr.L2.SAS}
\alias{read.hocr.L2.SAS}
\title{Reads HyperOCR data collected using Satlantic
      Hyperspectral Surface Aquisition System (HyperSAS)}
\arguments{
\item{fn}{is the file name}

\item{VERBOSE}{is a logical parameter indicating whether or
not the reading steps are printed. Default is TRUE.}

\item{APPLY.DARK}{is a logical parameter indicating whether or not
the darks are subtracted to the measurements.
Default is TRUE.}
}
\value{
It returns a large list with all the parameters extracted
from the L2 file.
}
\description{
The function reads ASCII files of Level 2 (calibrated) produced by ProSoft
software and returns a large list contaning all data.
Internal darks measurements may be subtracted or not using the
logical parameter APPLY.DARK.
}
\details{
The dark measurements recorded during the measurements are available
in the L2 file and could be apply to the measurements if requested.
The lat/lon position and time are extracted from the header file.
Sun elevation is calculated from those information and stored in the "dd"
data frame returned in the list.

(NOTE: Only *.dat file generated by Prosoft version 7.7.16 are supported)
}
\examples{


}
\seealso{

}
\author{
Simon Bélanger
}
