% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rrs2FU.R
\name{Rrs2FU}
\alias{Rrs2FU}
\title{Compute the Forel-Ule water class from a remote sensing reflectance measurement}
\arguments{
\item{Waves}{vector of wavelenghts for the input spectral reflectance spectra}

\item{Rrs}{vector of the input spectral reflectance spectra (same lenght as waves)}
}
\description{
Compute the Forel-Ule water class from a remote sensing reflectance measurement
}
\author{
Simon Bélanger
}
