% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.process.HOCR.batch.R
\name{run.process.HOCR.batch}
\alias{run.process.HOCR.batch}
\title{Run batch processing for HyperOCR}
\arguments{
\item{data.path}{is the full path where the L2 data are stored (*.dat)}

\item{Ag.path}{is the full path where the CDOM  absorption files (RData format) are stored}

\item{Ap.path}{is the full path where the CDOM files (RData format) are stored}
}
\description{
This programe will read the log file and process every file
foung in the log file. The files must be put in the same folders
(e.g., ../dat/). The output will be save as PNG and RData.
The user can provide IOPs to estimate the K_Lu or K_u in order to make
the extrapolation of the Lu0- measured near the sea surface
(typically around 10 to 15 cm from the sea surface).
}
\details{
No shadow correction is applied yet.


 @param log.file is the name of the ASCII file containing the
list of L2 data to processed with parameters
(see process.HOCR for details).
}
\author{
Simon Bélanger
}
