% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralRho.2.sensor.R
\name{spectralRho.2.sensor}
\alias{spectralRho.2.sensor}
\title{Compute reflectance (or radiance) at a sensor wavebands
from an hyperspectral or multispectral measurement using
relative spectral responses (RSR)}
\arguments{
\item{waves}{vector of wavelenghts for the input spectral reflectance spectra}

\item{rho}{vector of the input spectral reflectance spectra (same lenght as waves)}

\item{SENSOR}{is a string for the SENSOR names. Supported sensors are
OLI (default),
MERIS,
S2A,
S2B,
S3A,
S3B,
MODISA,
MODIST,
VIIRS,
SeaWIFS,
RapidEye,
PlanetScope_0C0D,
PlanetScope_0E,
PlanetScope_0F_10}
}
\description{
Compute reflectance (or radiance) at a sensor wavebands
from an hyperspectral or multispectral measurement using
relative spectral responses (RSR)
}
\author{
Simon Bélanger
}
