% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kd.Lee.R
\name{Kd.Lee}
\alias{Kd.Lee}
\title{Semi-analytical model of Lee et al JGR 2005.}
\usage{
Kd.Lee(a, bb)
}
\arguments{
\item{a}{is a vector of absorption coefficient}

\item{b}{os a vector of backscattering coefficient (length is the same as a)}
}
\description{
Semi-analytical model of Lee et al JGR 2005.
}
\author{
Simon Bélanger
}
