% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.HOCR.R
\name{process.HOCR}
\alias{process.HOCR}
\title{Processed in-water HyperOCR data}
\arguments{
\item{filen}{is the name of a L2 file in *.dat format procuded by
Prosoft}

\item{Station}{is the station ID}

\item{Lu0.Depth}{is the approximate depth of the Lu0 sensor
(Default is 0.1 m)}

\item{EdZ.Depth}{is the depth of the EdZ sensor
(Default is NA assuming there is no EdZ sensor)}

\item{Delta.LuZ.Depth}{is the distance between Lu0 and LuZ sensors
(Default is NA assuming there is no LuZ sensor)}

\item{Ag.file}{is the full file name and path of the RData file containing CDOM absorption}

\item{Ap.file}{is the full file name and path of the RData file containing particles absorption}

\item{Bbp.file}{is the full file name and path of the RData file containing bbp}
}
\description{
This routine compute AOPs from in-water hyperspectral
measurements performed using HyperOCR radiometers deployed
at the water surface with an arm and (optionaly) with a
tripod. Three set-ups have been implemented so far:
}
\details{
1. Upwelling radiance near the water surface (i.e. Lu0) only
2. Upwelling radiance near the water surface (i.e. Lu0) and
downwelling irradiance at depth Z (EdZ)
3. Upwelling radiance near the water surface (i.e. Lu0) and
upwelling radiance at depth Z (LuZ).
}
\author{
Simon Bélanger
}
