% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAA.v5.R
\name{QAA.v5}
\alias{QAA.v5}
\title{Quasi-Analytical Algorithm to retreive spectral absorption and backscattering coefficients.
The current paramterisation is version 5}
\usage{
QAA.v5(waves, Rrs)
}
\arguments{
\item{waves}{is a vector of wavelenghts}

\item{Rrs}{os a vector of Rrs (length is the same as waves)}
}
\description{
Quasi-Analytical Algorithm to retreive spectral absorption and backscattering coefficients.
The current paramterisation is version 5
}
\author{
Simon Bélanger
}
