% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.process.HOCR.batch.R
\name{run.process.HOCR.batch}
\alias{run.process.HOCR.batch}
\title{Run batch processing for surface in-water HyperOCR measurements}
\arguments{
\item{log.file}{is the name of the ASCII file containing the
list of L2 data to processed with parameters
(see process.HOCR for details).}

\item{data.path}{is the full path where the L2 data folder is stored (./dat/*.dat).
Note that the path should not include the "dat" folder name. On the parent folder is needed.}

\item{Ag.path}{is the full path where the CDOM  absorption files (RData format) are stored}

\item{Ap.path}{is the full path where the particulate absorption files (RData format) are stored}

\item{Bbp.path}{is the full path where the particulate backscattering files (RData format) are stored}
}
\description{
This programe reads a log file in which *L2.dat files coming for prosoft are listed for processing.
Each file will be processed using the parameters provided in the log file. The files must be put in the same folders
(e.g., ../dat/). The output will be save as PNG and RData.
The user can provide IOPs to estimate the K_Lu or K_u in order to make
the extrapolation of the Lu0- measured near the sea surface
(typically around 10 to 15 cm from the sea surface).
}
\details{
No shadow correction is applied yet.


The HOCR radiometers can be deployed from small boats to estime hyperspectral remote sensing reflectance.
Diffent setup are supported but the minimum is to heve one reference sensor in the air for Ed0+ and
at least one sensor in-water for either Eu, Lu or Ed.  The depth at which the in-water sensor is emerged
is given in the processing parameters in the log file. For example, if the Lu sensor near was lowered at 10cm below the
surface, Lu0.Depth = 0.1. If the set up uses a Eu0 sensor, then Eu0.Depth must be provided.
More accurate Rrs can be obtained using two Lu sensors emerged near the surface at two different depths,
which allow a direct assessment of the diffuse attenuation coefficient for upwelling radiance (K_Lu).

See \code{\link{process.HOCR}} for more details.
}
\author{
Simon Bélanger
}
