% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.process.HOCR.batch.R
\name{run.process.HOCR.batch}
\alias{run.process.HOCR.batch}
\title{Run batch processing for HyperOCR}
\usage{
run.process.HOCR.batch(log.file = "log.txt", data.path = "./",
  Ag.path = "./", Ap.path = "./", Bbp.path = "./")
}
\arguments{
\item{data.path}{is the full path where the L2 data are stored (*.dat)}
}
\description{
@param log.file is the name of the ASCII file containing the
list of L2 data to processed.
}
\author{
Simon Bélanger
}
